feat(data-loaders): allow `data` to always be defined
BREAKING CHANGE: Based on the `options` passed to a `defineLoader()`
function, the `data` will not be possibly `undefined`. This related to
[#319](https://github.com/posva/unplugin-vue-router/issues/319). To make
this possible, the way the `errors` option works has changed. The
[documentation](https://uvr.esm.is/data-loaders/error-handling.html#error-handling)
has been updated. Overall, the local options always take precedence now
and by default, no errors are considered as expected (`errors: false`).