refactor: add Error inference (will be removed)
As noted in the docs of this commit, the error can come from unexpected
places. Therefore, it's not correct to set its type. In the end, we will
have to check for the error type in a component anyway, so it's better
not to have the error type inference. It will be removed in the next
commit.