refactor: stage navigation result for checks
This makes me wonder if we're really returning a NavigationResult
because it does complexify the implementation and the checks. Most of
the time, when you want to redirect or change the navigation, you want
to change it. You just throw the NavigationResult, and then it just
works because it's an error and it propagates. We don't need the select
NavigationResult property either.

Maybe we need to revisit these before going stable.